// src/app/models/product.ts
export interface Product {
    uid?: string;
    id?: string;
    imageUrl?: string;
    name?: string;
    price?: number;
    delprice?: number;
    ProdcutCatId?: string;
    ProductCatName?: string;
    description?: string;
    isActive?: boolean;
    amount?: number;
    amountVent?: number;
    userUid?: string;
    userName?: string;
    images?: string;
    date?: string;
}